+++ src/mpi/romio/adio/common/shfp_fname.c
@@ -58,7 +58,7 @@ void ADIOI_Shfp_fname(ADIO_File fd, int rank)
 	sprintf(tmp, ".shfp.%d", i);
 	strcat(fd->shared_fp_fname, tmp);
 	
-	len = strlen(fd->shared_fp_fname);
+	len = (int)strlen(fd->shared_fp_fname);
 	MPI_Bcast(&len, 1, MPI_INT, 0, fd->comm);
 	MPI_Bcast(fd->shared_fp_fname, len+1, MPI_CHAR, 0, fd->comm);
     }
